% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_cmd_args}
\alias{collapse_cmd_args}
\title{Collapse Command Arguments}
\usage{
collapse_cmd_args(cmd, args)
}
\arguments{
\item{cmd}{Character. The command to be run, e.g. `plink2`}

\item{args}{Character. Vector of the arguments to `cmd`.}
}
\value{
A single character string that can be passed to the `-icmd` flag of
  the Swiss Army Knife
}
\description{
When submitting jobs to the Swiss Army Knife, this function is a convenience
function to collapse the function arguments into a single call before
submitting the call to [dx_run_swiss_army_knife()].
}
\examples{
\dontrun{
args <- "plink2"
cmd <- c("pfile", "ofh_data", "--make-bed", "--out", "ofh_data_bed")

collapse_cmd_args(args, cmd)
}
}
