% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bgen_annotation}
\alias{create_bgen_annotation}
\title{Create BGEN File Annotation}
\usage{
create_bgen_annotation(project_id, bgen_dir)
}
\arguments{
\item{project_id}{Character. An DNAnexus project ID where the data was
dispensed.}

\item{bgen_dir}{Character. The path to the raw BGEN files. Should be either
`/snv_bgen/` or `/imputed_bgen/`}
}
\value{
A data.table with the full paths to the BGEN files and some parsed
  information
}
\description{
This function tries to parse an OFH directory for BGEN files and creates an
annotation table that makes processing the data simpler.
}
