% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bgen_file_names}
\alias{create_bgen_file_names}
\title{Create OFH genomic file BGEN file names}
\usage{
create_bgen_file_names(file_prefix)
}
\arguments{
\item{file_prefix}{Character. Vector of file prefixes that the `.bgen` ending
will be applied to.}
}
\value{
BGEN file names
}
\description{
This function simply adds the `.bgen` ending to the supplied file prefix
}
