% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_genomic_file_batch_ids}
\alias{create_genomic_file_batch_ids}
\title{Create batch ID string for genomic files}
\usage{
create_genomic_file_batch_ids(n_batches)
}
\arguments{
\item{n_batches}{Integer. Number of batches for which IDs should be
generated. Enumerates from 1..`n_batches`}
}
\value{
Formatted character vector, e.g. c("b0001", "b0002") for for input
  `n_batches = 2`
}
\description{
Create formatted string for genomic file batches. Can be used to create
genomic file names.
}
