% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_genomic_file_prefix}
\alias{create_genomic_file_prefix}
\title{Create genomic file prefix}
\usage{
create_genomic_file_prefix(prefix, version, chr, n_batches)
}
\arguments{
\item{prefix}{Character. [glue::glue()] style string for formatting, e.g.
"ofh_imputed.\{version\}.\{chr\}-\{batch\}". Needs to contain `version`,
`chr` & `batch`}

\item{version}{Character. The release version of the genomic data, e.g. `v9`}

\item{chr}{Character. The chromosome identifier, Ensembl style: chr1, chr2, ... chrX}

\item{n_batches}{Integer. Number of batches for which IDs should be
generated. Enumerates from 1..`n_batches`}
}
\value{
Formatted prefixes
}
\description{
Create a formatted string containing chromosome and batch ID that can be used to
}
