% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_imputed_file_prefix}
\alias{create_imputed_file_prefix}
\title{Create OFH imputed genomic file prefix}
\usage{
create_imputed_file_prefix(chr, n_batches)
}
\arguments{
\item{chr}{Character. The chromosome identifier, Ensembl style: chr1, chr2, ... chrX}

\item{n_batches}{Integer. Number of batches for which IDs should be
generated. Enumerates from 1..`n_batches`}
}
\value{
Formatted character strings.
}
\description{
Create OFH imputed genomic file prefix
}
