% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_sample_file_names}
\alias{create_sample_file_names}
\title{Create OFH imputed genomic file BGEN Sample file names}
\usage{
create_sample_file_names(file_prefix)
}
\arguments{
\item{file_prefix}{Character. Vector of file prefixes that the `.sample`
ending will be applied to.}
}
\value{
BGEN Sample file names
}
\description{
Create OFH imputed genomic file BGEN Sample file names
}
