% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decompress_gzip}
\alias{decompress_gzip}
\title{Decompress gzipped files}
\usage{
decompress_gzip(files)
}
\arguments{
\item{files}{gzipped files to be decompressed}
}
\value{
TRUE on success
}
\description{
Decompress gzipped files using the `gzip` utility
}
