% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_clear_env}
\alias{dx_clear_env}
\title{Clear DNAnexus Environment}
\usage{
dx_clear_env()
}
\value{
Invisible TRUE if successful
}
\description{
Clears the current DNAnexus environment settings.
}
\examples{
\dontrun{
# Clear environment
# dx_clear_env()
}
}
