% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{dx_get_path}
\alias{dx_get_path}
\title{Get the DNAnexus path}
\usage{
dx_get_path()
}
\value{
Character string containing the DNAnexus path
}
\description{
Retrieves the currently set DNAnexus path from the cache after validating
that a path is set.
}
