% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{dx_get_project}
\alias{dx_get_project}
\title{Get the DNAnexus project ID}
\usage{
dx_get_project()
}
\value{
Character string containing the DNAnexus project ID
}
\description{
Retrieves the currently set DNAnexus project ID from the cache after validating
that a project is set.
}
