% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_parse_env}
\alias{dx_parse_env}
\title{Parse DNAnexus-specific Environmental Variables}
\usage{
dx_parse_env()
}
\value{
Character vector containing relevant variables formatted as `VARIABLE=VALUE`
}
\description{
This is a convenience function that parses the environmental variables
returned by `dx env` for use with [dx_run_cmd].
}
