% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_remount_project}
\alias{dx_remount_project}
\title{Remount DNAnexus Project}
\usage{
dx_remount_project()
}
\value{
Invisible TRUE if successful
}
\description{
Remounts the DNAnexus project file system.
}
\examples{
\dontrun{
# Remount project
# dx_remount_project()
}
}
