% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_set_binary}
\alias{dx_set_binary}
\title{Set DNAnexus Binary Path}
\usage{
dx_set_binary(dx_binary = NULL, ...)
}
\arguments{
\item{dx_binary}{Character string specifying the path to the dx binary.
If NULL (default), assumes "dx" is in PATH.}

\item{...}{parameters to pass to `dx_run_cmd`}
}
\value{
Invisible TRUE if successful
}
\description{
Sets the path to the DNAnexus dx binary. This is typically called internally
by dx_init().
}
\examples{
\dontrun{
# Set dx binary path
# dx_set_binary("/usr/local/bin/dx")
}
}
