% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_set_env}
\alias{dx_set_env}
\title{Set DNAnexus Environment Cache}
\usage{
dx_set_env()
}
\value{
Invisible TRUE if successful
}
\description{
Sets the internal cache with current DNAnexus environment information.
This is typically called internally by dx_init().
}
\examples{
\dontrun{
# Set environment cache
# dx_set_env()
}
}
