% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_set_project}
\alias{dx_set_project}
\title{Set DNAnexus Project}
\usage{
dx_set_project(dx_project_id = NULL, ...)
}
\arguments{
\item{dx_project_id}{Character string specifying the DNAnexus project ID.}

\item{...}{parameters to pass to `dx_run_cmd`}
}
\value{
Invisible TRUE if successful, aborts on failure
}
\description{
Selects a DNAnexus project to work with.
}
\examples{
\dontrun{
# Set project
# dx_set_project("project-xxxxx")
}
}
