% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{get_dx_cache}
\alias{get_dx_cache}
\title{Get values from the DNAnexus cache}
\usage{
get_dx_cache(dx_property = "all")
}
\arguments{
\item{dx_property}{Character string specifying which property to retrieve.
Use "all" to retrieve all properties. Default is "all".}
}
\value{
Value(s) from the cache environment
}
\description{
Retrieves one or more values from the DNAnexus cache environment.
}
