% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rate_card}
\alias{rate_card}
\title{Our Future Health TRE rate card}
\format{
## `rate_card`
A data frame with 48 rows and 9 columns:
\describe{
  \item{dnanexus_instance_type}{Character string with the DNAnexus instance type identifier}
  \item{azure_instance_type}{Character string with the Azure VM instance type}
  \item{n_gpus}{Integer number of GPUs}
  \item{n_cpus}{Integer number of CPU cores}
  \item{ram_gb}{Integer RAM in gigabytes}
  \item{disk_storage_gb}{Integer disk storage in gigabytes}
  \item{storage_type}{Character string describing storage type}
  \item{on_demand_rate_gpb_per_h}{Numeric on-demand rate in GBP per hour}
  \item{spot_rate_gpb_per_h}{Numeric spot rate in GBP per hour}
}
}
\source{
<https://a.storyblok.com/f/228028/x/bf1f0d1c41/our_future_health_trusted_research_environment_rate_card.pdf>
}
\usage{
rate_card
}
\description{
The rate card for cloud computing instances used when working on the DNAnexus
cloud environment, specifically for the Our Future Health study, which
differs from default instances.
}
\keyword{datasets}
