% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{remove_dx_cache}
\alias{remove_dx_cache}
\title{Remove all entries from the DNAnexus cache}
\usage{
remove_dx_cache()
}
\value{
Invisible NULL
}
\description{
Removes all named entries from the DNAnexus cache environment.
}
