% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_tre_proxy_env_var}
\alias{remove_tre_proxy_env_var}
\title{Remove proxy environment variable on OFH TRE}
\usage{
remove_tre_proxy_env_var()
}
\value{
Character vector containing relevant variables formatted as `VARIABLE=VALUE`
}
\description{
This is a convenience function that removes the environmental for the https
proxy when a shell command is run within an R session on the OFH DNAnexus TRE.
This is done to mitigate an error when running `dx` commands that seems to be
caused by this variable.
}
