% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{reset_dx_cache}
\alias{reset_dx_cache}
\title{Reset the DNAnexus cache}
\usage{
reset_dx_cache()
}
\description{
Clears all stored DNAnexus configuration values, setting them to NULL
except for the initialized flag which is set to FALSE.
}
