% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{set_dx_cache}
\alias{set_dx_cache}
\title{Set values in the DNAnexus cache}
\usage{
set_dx_cache(...)
}
\arguments{
\item{...}{Named arguments to set in the cache}
}
\value{
Invisible NULL
}
\description{
Assigns one or more values to the DNAnexus cache environment.
}
