% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tre_path}
\alias{tre_path}
\title{Create TRE-compatible path}
\usage{
tre_path(project, ...)
}
\arguments{
\item{project}{Character. The TRE project ID where the path is located.}

\item{...}{Character. Character vectors that will be passed to [fs::path()]
that make up the full path.}
}
\value{
Character. A single character string describing a path in the form
  "project:/path/to/file"
}
\description{
This is a thin wrapper around [fs::path()] that creates a full path
consisting of the project ID and the full path by appending a colon `:` to
the project ID before adding the path, e.g. `project:/path/to/file`
}
\examples{
tre_path("project-12345", "path", "to", "some", "file")

}
