% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bgen_file_names}
\alias{create_bgen_file_names}
\title{Create OFH imputed genomic file BGEN file names}
\usage{
create_bgen_file_names(chr, n_batches)
}
\arguments{
\item{chr}{Character. The chromosome identifier, Ensembl style: chr1, chr2, ... chrX}

\item{n_batches}{Integer. Number of batches for which IDs should be
generated. Enumerates from 1..`n_batches`}
}
\value{
OFH imputed BGEN file names
}
\description{
Create OFH imputed genomic file BGEN file names
}
