% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_utils.R
\name{dx_set_path}
\alias{dx_set_path}
\title{Set DNAnexus Path}
\usage{
dx_set_path(dx_path = NULL)
}
\arguments{
\item{dx_path}{Character string specifying the DNAnexus project path.}
}
\value{
Invisible TRUE if successful, aborts on failure
}
\description{
Sets the working directory within the DNAnexus project.
}
\examples{
\dontrun{
# Set path
# dx_set_path("/data/subfolder")
}
}
