% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cache.R
\name{init_dx_cache}
\alias{init_dx_cache}
\title{Initialize the DNAnexus cache environment}
\usage{
init_dx_cache()
}
\value{
An environment with slots for storing DNAnexus configuration
}
\description{
Creates a new environment to store DNAnexus configuration values.
}
